// app-web/vela-wallet/src/lib/core/generated/SignPrefOperation.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SignPrefOperation = { "type": "read_stored" } | { "type": "write_method", method: string, } | { "type": "write_signer_url", url: string | null, } | { "type": "write_relay_url", url: string | null, };